move add_tlc and remove_tlc to dev module
