+++ src/rpc/channel.rs
@@ -1,15 +1,13 @@
 use crate::fiber::{
     channel::{
-        AddTlcCommand, AwaitingChannelReadyFlags, AwaitingTxSignaturesFlags,
-        ChannelActorStateStore, ChannelCommand, ChannelCommandWithId,
-        ChannelState as RawChannelState, CloseFlags, CollaboratingFundingTxFlags,
-        NegotiatingFundingFlags, RemoveTlcCommand, ShutdownCommand, ShuttingDownFlags,
+        AwaitingChannelReadyFlags, AwaitingTxSignaturesFlags, ChannelActorStateStore,
+        ChannelCommand, ChannelCommandWithId, ChannelState as RawChannelState, CloseFlags,
+        CollaboratingFundingTxFlags, NegotiatingFundingFlags, ShutdownCommand, ShuttingDownFlags,
         SigningCommitmentFlags, UpdateCommand,
     },
-    hash_algorithm::HashAlgorithm,
     network::{AcceptChannelCommand, OpenChannelCommand},
     serde_utils::{EntityHex, U128Hex, U64Hex},
-    types::{Hash256, RemoveTlcFulfill, TlcErr, TlcErrPacket, TlcErrorCode, NO_SHARED_SECRET},
+    types::Hash256,
     NetworkActorCommand, NetworkActorMessage,
 };
 use crate::{handle_actor_call, handle_actor_cast, log_and_error};
@@ -29,7 +27,6 @@ use ractor::{call, ActorRef};
 use serde::{Deserialize, Serialize};
 use serde_with::{serde_as, DisplayFromStr};
 use std::cmp::Reverse;
-use std::str::FromStr;
 use tentacle::secio::PeerId;
 
 #[serde_as]
@@ -257,54 +254,6 @@ pub(crate) struct Channel {
 }
 
 #[serde_as]
-#[derive(Serialize, Deserialize, Debug)]
-pub(crate) struct AddTlcParams {
-    /// The channel ID of the channel to add the TLC to
-    channel_id: Hash256,
-    /// The amount of the TLC
-    #[serde_as(as = "U128Hex")]
-    amount: u128,
-    /// The payment hash of the TLC
-    payment_hash: Hash256,
-    /// The expiry of the TLC
-    #[serde_as(as = "U64Hex")]
-    expiry: u64,
-    /// The hash algorithm of the TLC
-    hash_algorithm: Option<HashAlgorithm>,
-}
-
-#[serde_as]
-#[derive(Clone, Serialize)]
-pub(crate) struct AddTlcResult {
-    /// The ID of the TLC
-    #[serde_as(as = "U64Hex")]
-    tlc_id: u64,
-}
-
-#[serde_as]
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub(crate) struct RemoveTlcParams {
-    /// The channel ID of the channel to remove the TLC from
-    channel_id: Hash256,
-    #[serde_as(as = "U64Hex")]
-    /// The ID of the TLC to remove
-    tlc_id: u64,
-    /// The reason for removing the TLC, either a 32-byte hash for preimage fulfillment or an u32 error code for removal
-    reason: RemoveTlcReason,
-}
-
-/// The reason for removing a TLC
-#[serde_as]
-#[derive(Serialize, Deserialize, Debug, Clone)]
-#[serde(untagged)]
-enum RemoveTlcReason {
-    /// The reason for removing the TLC is that it was fulfilled
-    RemoveTlcFulfill { payment_preimage: Hash256 },
-    /// The reason for removing the TLC is that it failed
-    RemoveTlcFail { error_code: String },
-}
-
-#[serde_as]
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub(crate) struct ShutdownChannelParams {
     /// The channel ID of the channel to shut down
@@ -367,14 +316,6 @@ trait ChannelRpc {
         &self,
         params: CommitmentSignedParams,
     ) -> Result<(), ErrorObjectOwned>;
-
-    /// Adds a TLC to a channel.
-    #[method(name = "add_tlc")]
-    async fn add_tlc(&self, params: AddTlcParams) -> Result<AddTlcResult, ErrorObjectOwned>;
-
-    /// Removes a TLC from a channel.
-    #[method(name = "remove_tlc")]
-    async fn remove_tlc(&self, params: RemoveTlcParams) -> Result<(), ErrorObjectOwned>;
 
     /// Shuts down a channel.
     #[method(name = "shutdown_channel")]
@@ -514,78 +455,6 @@ where
             },
         ));
         handle_actor_cast!(self.actor, message, params)
-    }
-
-    async fn add_tlc(&self, params: AddTlcParams) -> Result<AddTlcResult, ErrorObjectOwned> {
-        let message = |rpc_reply| -> NetworkActorMessage {
-            NetworkActorMessage::Command(NetworkActorCommand::ControlFiberChannel(
-                ChannelCommandWithId {
-                    channel_id: params.channel_id,
-                    command: ChannelCommand::AddTlc(
-                        AddTlcCommand {
-                            amount: params.amount,
-                            payment_hash: params.payment_hash,
-                            expiry: params.expiry,
-                            hash_algorithm: params.hash_algorithm.unwrap_or_default(),
-                            onion_packet: None,
-                            shared_secret: NO_SHARED_SECRET.clone(),
-                            previous_tlc: None,
-                        },
-                        rpc_reply,
-                    ),
-                },
-            ))
-        };
-        handle_actor_call!(self.actor, message, params).map(|response| AddTlcResult {
-            tlc_id: response.tlc_id,
-        })
-    }
-
-    async fn remove_tlc(&self, params: RemoveTlcParams) -> Result<(), ErrorObjectOwned> {
-        let err_code = match &params.reason {
-            RemoveTlcReason::RemoveTlcFail { error_code } => {
-                let Ok(err) = TlcErrorCode::from_str(&error_code) else {
-                    return log_and_error!(params, format!("invalid error code: {}", error_code));
-                };
-                Some(err)
-            }
-            _ => None,
-        };
-        let message = |rpc_reply| -> NetworkActorMessage {
-            NetworkActorMessage::Command(NetworkActorCommand::ControlFiberChannel(
-                ChannelCommandWithId {
-                    channel_id: params.channel_id,
-                    command: ChannelCommand::RemoveTlc(
-                        RemoveTlcCommand {
-                            id: params.tlc_id,
-                            reason: match &params.reason {
-                                RemoveTlcReason::RemoveTlcFulfill { payment_preimage } => {
-                                    crate::fiber::types::RemoveTlcReason::RemoveTlcFulfill(
-                                        RemoveTlcFulfill {
-                                            payment_preimage: *payment_preimage,
-                                        },
-                                    )
-                                }
-                                RemoveTlcReason::RemoveTlcFail { .. } => {
-                                    // TODO: maybe we should remove this PRC or move add_tlc and remove_tlc to `test` module?
-                                    crate::fiber::types::RemoveTlcReason::RemoveTlcFail(
-                                        TlcErrPacket::new(
-                                            TlcErr::new(err_code.expect("expect error code")),
-                                            // Do not encrypt the error message when removing the TLC via RPC.
-                                            // TODO: use tlc id to look up the shared secret in the store
-                                            &NO_SHARED_SECRET,
-                                        ),
-                                    )
-                                }
-                            },
-                        },
-                        rpc_reply,
-                    ),
-                },
-            ))
-        };
-
-        handle_actor_call!(self.actor, message, params)
     }
 
     async fn shutdown_channel(
